Temporarily move notification banner files to check duplications
