+++ src/main/steps/common/task-list/utils.ts
@@ -7,7 +7,7 @@ import { UserDetails } from '../../../app/controller/AppRequest';
 import { getPartyDetails } from '../../../steps/tasklistresponse/utils';
 
 import { CaseType, PartyDetails, PartyType, Respondent, State, YesOrNo } from './../../../app/case/definition';
-import { findC7ResponseDocument } from './components/notification-banner/utils';
+import { findC7ResponseDocument } from './components/notification-banner-temp/utils';
 
 export const getPartyName = (
   caseData: Partial<CaseWithId> | undefined,
